Added an option to not copy the Meteor docs server when building the docs.
Setting "initMeteor" to false allows to customize the Meteor docs server after the first build without seeing the next build overriding the changes.
When using "meteor-jsdoc build" for the first time, it requires the "initMeteor" setting to be true, otherwise, only the data files will be copied, and you won't be able to start the server (there will be none).
Updated readme.